refactor(client): refactor client util library
- [x] clean up client-util sources (remove unneeded sources, rearrange directory structure);
