+++ libs/client-diagnostics/src/lib/components/home/diagnostics-home.component.ts
@@ -1,10 +1,14 @@
 import { ChangeDetectionStrategy, Component } from '@angular/core';
 import { AppMarkdownService } from '@app/client-services';
-import { TIMEOUT } from '@app/client-util';
 import { of, timer } from 'rxjs';
 import { first, map, take } from 'rxjs/operators';
 
 const counter = 5;
+
+const timeout = {
+  start: 0,
+  interval: 2000,
+};
 
 @Component({
   selector: 'app-diagnostics-home',
@@ -13,7 +17,7 @@ const counter = 5;
   changeDetection: ChangeDetectionStrategy.OnPush,
 })
 export class AppDiagnosticsHomeComponent {
-  public readonly timer$ = timer(TIMEOUT.INSTANT, TIMEOUT.MEDIUM).pipe(
+  public readonly timer$ = timer(timeout.start, timeout.interval).pipe(
     take(counter),
     map(num => `Until destroyed ${num}`),
   );
