+++ libs/client-unit-testing/src/lib/jest/jest-jsdom-globals.patch.ts
@@ -1,7 +1,5 @@
 // import 'node_modules/jsdom/lib/jsdom/living/custom-elements/CustomElementRegistry-impl.js';
 
-import { TIMEOUT } from '@app/client-util';
-
 import { setupJsdomDocumentMocks } from './document/document.mock';
 import { setupJsdomGlobalMocks } from './globals/globals.mock';
 import { setupJsdomWindowMocks } from './window/window.mock';
@@ -9,7 +7,8 @@ import { setupJsdomWindowMocks } from './window/window.mock';
 /**
  * Increase specs timeout.
  */
-jest.setTimeout(TIMEOUT.FOREVER);
+const timeout = 10000;
+jest.setTimeout(timeout);
 
 /**
  * Sets up Jest global mocks
