+++ libs/client-util/src/index.ts
@@ -1,8 +1,5 @@
-export * from './lib/animations/index';
-export * from './lib/decorators/index';
-export * from './lib/general-purpose/index';
-export * from './lib/general-purpose/index';
-export * from './lib/http/enums';
-export * from './lib/interfaces/index';
-export * from './lib/ngxs/index';
+export * from './lib/animations';
+export * from './lib/decorators';
+export * from './lib/interfaces';
+export * from './lib/providers';
 export * from './lib/security';
