--- libs/client-util/src/lib/general-purpose/enums.ts
@@ -1,9 +1,0 @@
-export enum TIMEOUT {
-  INSTANT = 0,
-  SHORTEST = 500,
-  SHORT = 1000,
-  MEDIUM = 2000,
-  LONG = 3500,
-  LONGEST = 5000,
-  FOREVER = 10000,
-}
