--- libs/client-util/src/lib/general-purpose/factories.ts
@@ -1,13 +1,0 @@
-/**
- * Window factory.
- */
-export function windowFactory() {
-  return window;
-}
-
-/**
- * Document factory.
- */
-export function documentFactory() {
-  return window.document;
-}
