--- libs/client-util/src/lib/general-purpose/index.ts
@@ -1,4 +1,0 @@
-export * from './enums';
-export * from './factories';
-export * from './injection-tokens';
-export * from './providers';
