+++ libs/client-util/src/lib/interfaces/index.ts
@@ -1,4 +1,6 @@
 export * from './environment.interface';
+export * from './http-status.interface';
+export * from './ngxs-action-payload.interface';
 export * from './sentry.interface';
 export * from './toaster.interface';
 export * from './toolbar.interface';
