+++ libs/client-util/src/lib/interfaces/ngxs-action-payload.interface.ts
