--- libs/client-util/src/lib/ngxs/index.ts
@@ -1,1 +1,0 @@
-export * from './ngxs-action-payload.interface';
