+++ libs/client-util/src/lib/providers/document.provider.ts
@@ -1,0 +1,11 @@
+import { DOCUMENT } from '@angular/common';
+import { Provider } from '@angular/core';
+
+export function documentFactory() {
+  return window.document;
+}
+
+export const documentProvider: Provider = {
+  provide: DOCUMENT,
+  useFactory: documentFactory,
+};
