+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "nx-ng-starter",
-  "version": "1.2.0",
+  "version": "1.2.1",
   "private": true,
   "description": "Monorepo starter: Nx, Angular, Angular Elements, Electron, NodeJS, NestJS, Firebase.",
   "license": "MIT",
